youtube-transcript-summarizer-api/requirements.txt
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
Flask==2.0.1
Flask-Cors==3.0.10
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.3
numpy==1.21.2
packaging==21.0
pathy==0.6.0
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
regex==2021.9.30
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.1
youtube-transcript-api==0.4.1
moviepy==1.0.3
pytube==12.1.0
google-cloud-speech==2.16.2
google-cloud-storage==2.5.0